Backend/requirements.txt.txt
fastapi==0.95.0
uvicorn==0.21.0
sqlalchemy==2.0.0
alembic==1.10.2
python-dotenv==1.0.0
requests==2.28.2
pycardano==0.8.0
pydantic==1.10.6
psycopg2-binary==2.9.5
aiohttp==3.8.4
cbor2==5.4.6
blockfrost-python==0.5.0